requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
# mysqlclient  
PyMySQL      
pydantic[email]
python-dotenv
python-jose[cryptography]
pydantic-settings
passlib==1.7.4  
bcrypt==4.0.1